+++ BookStore/BookStore/Controllers/CategoryController.cs
@@ -25,6 +25,12 @@ namespace BookStore.Controllers
         [HttpGet]
         public IActionResult Add()
         {
+            
+            if (!IsAdmin(User))
+            {
+                return Unauthorized();
+            }
+
             CategoryViewModel model = new CategoryViewModel();
 
             return View(model);
@@ -33,6 +39,11 @@ namespace BookStore.Controllers
         [HttpPost]
         public async Task<IActionResult> Add(CategoryViewModel model)
         {
+            if (!IsAdmin(User))
+            {
+                return Unauthorized();
+            }
+
             await categoryService.AddCategoryAsync(model);
 
             TempData["Success"] = "Category added succesfully!";
@@ -42,6 +53,11 @@ namespace BookStore.Controllers
 
         public async Task<IActionResult> Remove(int id)
         {
+            if (!IsAdmin(User))
+            {
+                return Unauthorized();
+            }
+
             await categoryService.RemoveAsync(id);
 
             TempData["Success"] = "Category removed succesfully!";
@@ -52,6 +68,11 @@ namespace BookStore.Controllers
         [HttpGet]
         public IActionResult Edit(int id)
         {
+            if (!IsAdmin(User))
+            {
+                return Unauthorized();
+            }
+
             Category model = categoryService.FindCategory(id);
 
             return View(model);
@@ -59,6 +80,12 @@ namespace BookStore.Controllers
 
         public async Task<IActionResult> Edit(int id, Category model)
         {
+            if (!IsAdmin(User))
+            {
+                return Unauthorized();
+            }
+
+
             model.Id = id;
 
             await categoryService.EditCategoryAsync(model);
